+++ src/main/java/com/diffplug/gradle/pde/EclipseRelease.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2015-2023 DiffPlug
+ * Copyright (C) 2015-2024 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -58,7 +58,7 @@ public final class EclipseRelease {
 		}
 	}
 
-	public static final String LATEST = "4.30.0";
+	public static final String LATEST = "4.32.0";
 
 	public static EclipseRelease latestOfficial() {
 		return official(LATEST);
@@ -132,6 +132,8 @@ public final class EclipseRelease {
 			case "4.28.0": return root + "4.28/R-4.28-202306050440/";
 			case "4.29.0": return root + "4.29/R-4.29-202309031000/";
 			case "4.30.0": return root + "4.30/R-4.30-202312010110/";
+			case "4.31.0": return root + "4.31/R-4.31-202402290520/";
+			case "4.32.0": return root + "4.32/R-4.32-202406010610/";
 			// less-specific versions
 			case "3.5": case "3.6": case "3.7": case "3.8":
 			case "4.2": case "4.3": case "4.4": case "4.5":
@@ -141,7 +143,7 @@ public final class EclipseRelease {
 			case "4.18": case "4.19": case "4.20": case "4.21":
 			case "4.22": case "4.23": case "4.24": case "4.25":
 			case "4.26": case "4.27": case "4.28": case "4.29":
-			case "4.30":
+			case "4.30": case "4.31": case "4.32":
 				return root + v + "/";
 			default: return null;
 			}